Moved setBaseUrl function to seperate file
