+++ src/config/setBaseUrl.js
@@ -1,0 +1,14 @@
+const setBaseUrl = () => {
+  switch (process.env.NODE_ENV) {
+    case 'production':
+      return process.env.REACT_APP_API_BASE_URL_PROD;
+
+    case 'test':
+      return process.env.REACT_APP_API_BASE_URL_TEST;
+
+    default:
+      return process.env.REACT_APP_API_BASE_URL_DEV;
+  }
+};
+
+export default setBaseUrl;
